refactor: reorganizacion de codigo
